+++ src/app/trainer/home/page.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import { HomePage } from '@/pages/trainer/ui';
+
+const Home = () => {
+  return <HomePage />;
+};
+
+export default Home;
