+++ src/pages/member/ui/HomePage.tsx
@@ -1,0 +1,11 @@
+import MemberLayout from './MemberLayout';
+
+const HomePage = () => {
+  return (
+    <MemberLayout>
+      <div>member HomePage</div>
+    </MemberLayout>
+  );
+};
+
+export { HomePage };
