+++ src/pages/member/ui/MemberLayout.tsx
@@ -1,6 +1,15 @@
 'use client';
 
+import styled from 'styled-components';
+
 import { MemberNavigationBar } from '@/shared/ui/navigationBar';
+
+const MemberWrap = styled.section`
+  width: 360px;
+  height: 100vh;
+  margin: 0 auto;
+  background-color: #70909e;
+`;
 
 interface ComponentProps {
   children: React.ReactNode;
@@ -8,10 +17,10 @@ interface ComponentProps {
 
 const MemberLayout = ({ children }: ComponentProps) => {
   return (
-    <div>
+    <MemberWrap>
       {children}
       <MemberNavigationBar />
-    </div>
+    </MemberWrap>
   );
 };
