+++ src/pages/trainer/ui/HomePage.tsx
@@ -1,0 +1,11 @@
+import TrainerLayout from './TrainerLayout';
+
+const HomePage = () => {
+  return (
+    <TrainerLayout>
+      <div>trainer HomePage</div>
+    </TrainerLayout>
+  );
+};
+
+export { HomePage };
