+++ src/pages/trainer/ui/TrainerLayout.tsx
@@ -1,6 +1,15 @@
 'use client';
 
+import styled from 'styled-components';
+
 import { TrainerNavigationBar } from '@/shared/ui/navigationBar';
+
+const TrainerWrap = styled.section`
+  width: 360px;
+  height: 100vh;
+  margin: 0 auto;
+  background-color: #70909e;
+`;
 
 interface ComponentProps {
   children: React.ReactNode;
@@ -8,10 +17,10 @@ interface ComponentProps {
 
 const TrainerLayout = ({ children }: ComponentProps) => {
   return (
-    <div>
+    <TrainerWrap>
       {children}
       <TrainerNavigationBar />
-    </div>
+    </TrainerWrap>
   );
 };
