+++ src/pages/trainer/ui/index.ts
@@ -1,0 +1,1 @@
+export { HomePage } from './HomePage';
