--- src/shared/hooks/use-callback-ref/index.ts
@@ -1,1 +1,0 @@
-export { useCallbackRef } from './useCallbackRef';
