--- src/shared/hooks/use-controllable-state/index.ts
@@ -1,1 +1,0 @@
-export { useControllableState } from './useControllableState';
