+++ src/shared/ui/input/PasswordInput.tsx
@@ -14,6 +14,7 @@ const StyledPasswordWrap = styled.div`
 
 const Input = styled.input`
   width: 100%;
+  height: 100%;
 `;
 
 const StyledButton = styled(Button)`
