+++ src/shared/ui/input/index.ts
@@ -1,0 +1,3 @@
+export { EmailInput } from './EmailInput';
+export { PasswordInput } from './PasswordInput';
+export { TextInput } from './TextInput';
