--- src/shared/ui/input/passwordInput/index.ts
@@ -1,1 +1,0 @@
-export { PasswordInput } from './PasswordInput';
