--- src/shared/ui/input/phoneInput/index.ts
@@ -1,1 +1,0 @@
-export { EmailInput } from './EmailInput';
