--- src/shared/ui/input/textInput/index.ts
@@ -1,1 +1,0 @@
-export { TextInput } from './TextInput';
